Basic RPC implementation done.
* Fixed RPC_defs.h macro
* Better simple_comms
* Source::reset will not only reset cur_ind = 0 but also ACTIVE-ate the Source
* Old example programs are broken. wave_teensy and wave2_teensy and rpc_loop work.

+++ src/teensy/wave_teensy/wave_teensy.ino
@@ -5,7 +5,7 @@ uint16_t triangle_wave(uint8_t seq){
   return (abs(20 - seq)/40.0)*4095;
 }
 
-double triangle_update_period(uint8_t seq){
+uint32_t triangle_update_period(uint8_t seq){
   return 150000;
 }
